% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as_tibble}
\alias{as_tibble}
\title{Coerce an ldap3 response to a data frames}
\usage{
as_tibble(x, ...)
}
\arguments{
\item{x}{object of type "ldap3_response"}

\item{...}{Unused, for extensibility.}
}
\description{
Coerce an ldap3 response to a data frames
}
