% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{save_thumbnail_photo}
\alias{save_thumbnail_photo}
\title{Save thumbnailPhoto}
\usage{
save_thumbnail_photo(x, filename = "thumbnailPhoto.jpg")
}
\arguments{
\item{x}{raw bytes to save}

\item{filename}{name of the file}
}
\description{
Save thumbnailPhoto
}
\examples{
\dontrun{
conn <- wudap_connect()
response <- wudap_search("(mail=schuelke@wustl.edu)")
data <- as_tibble(response)
save_thumbnail_photo(data$thumbnailPhoto[[1]])
}
}
