% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wudap.R
\name{wudap_connect}
\alias{wudap_connect}
\title{Connect to WashU LDAP Server}
\usage{
wudap_connect(
  washu_key_user = sprintf("\%s@wustl.edu", Sys.getenv("WASHU_KEY_USER")),
  washu_key_pass = Sys.getenv("WASHU_KEY_PASS")
)
}
\arguments{
\item{washu_key_user}{WashU Key User}

\item{washu_key_pass}{WashU Key Password}
}
\value{
object of type "ldap3.core.connection.Connection"
}
\description{
Connect to WashU LDAP Server
}
\examples{
\dontrun{
conn <- wudap_connect()
}
}
