% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wudap.R
\name{wudap_quick_query}
\alias{wudap_quick_query}
\title{Washington University LDAP quick query}
\usage{
wudap_quick_query(
  type = c("sAMAccountName", "userPrincipalName", "mail"),
  value,
  user = sprintf("\%s@wustl.edu", Sys.getenv("WUSTL_KEY_USER")),
  pass = Sys.getenv("WUSTL_KEY_PASS"),
  separate = FALSE
)
}
\arguments{
\item{type}{the type of filter to construct}

\item{value}{the value for which to filter}

\item{user}{WUSTL Key username}

\item{pass}{WUSTL Key password}

\item{separate}{logical to detect and expand recursive columns}
}
\value{
curl response object
}
\description{
Perform generic LDAP queries
}
\note{
Must use campus IP address.
}
\examples{
\dontrun{
wudap_quick_query("sAMAccountName", "schuelke")

wudap_quick_query("userPrincipalName", "schuelke@wustl.edu")

wudap_quick_query("mail", "schuelke@wustl.edu")
}

}
