% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wudap.R
\name{wudap_quick_query_mail}
\alias{wudap_quick_query_mail}
\title{Washington University LDAP quick query}
\usage{
wudap_quick_query_mail(mail, user, pass)
}
\arguments{
\item{mail}{email address}

\item{user}{WUSTL Key username}

\item{pass}{WUSTL Key password}
}
\value{
curl response object
}
\description{
Perform generic LDAP queries
}
\note{
Must use campus IP address.
}
\examples{
\dontrun{
wudap_quick_query("sAMAccountName", "schuelke")

wudap_quick_query("userPrincipalName", "schuelke@wustl.edu")

wudap_quick_query("mail", "schuelke@wustl.edu")
}

}
