% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wudap.R
\name{wudap_quick_query_sAMAccountName}
\alias{wudap_quick_query_sAMAccountName}
\title{Washington University LDAP quick query}
\usage{
wudap_quick_query_sAMAccountName(sAMAccountName, user, pass)
}
\arguments{
\item{sAMAccountName}{a logon name used to support clients and servers from
previous version of Windows, such as Windows NT 4.0, Windows 95, Windows 98,
and LAN Manager}

\item{user}{WUSTL Key username}

\item{pass}{WUSTL Key password}
}
\value{
curl response object
}
\description{
Perform generic LDAP queries
}
\note{
Must use campus IP address.
}
\examples{
\dontrun{
wudap_quick_query("sAMAccountName", "schuelke")

wudap_quick_query("userPrincipalName", "schuelke@wustl.edu")

wudap_quick_query("mail", "schuelke@wustl.edu")
}

}
