% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wudap.R
\name{wudap_connect}
\alias{wudap_connect}
\title{Connect to WashU LDAP Server}
\usage{
wudap_connect(
  washu_key_id = sprintf("\%s@wustl.edu", Sys.getenv("WASHU_KEY_ID")),
  washu_key_password = Sys.getenv("WASHU_KEY_PASSWORD")
)
}
\arguments{
\item{user}{WashU Key ID}

\item{password}{WashU Key password}
}
\value{
object of type "ldap3.core.connection.Connection"
}
\description{
Connect to WashU LDAP Server
}
\examples{
\dontrun{
conn <- wudap_connect()
}
}
